concept_sliders/cs_requirements.txt
# bitsandbytes==0.41.1
# dadaptation==3.1
# ipython==8.7.0
# lion_pytorch==0.1.2
# lpips==0.1.4
# numpy==1.23.5
# opencv_python_headless==4.7.0.68
# pandas==1.5.2
# Pillow==10.1.0
# prodigyopt==1.0
# pydantic==2.6.3
# PyYAML==6.0.1
# Requests==2.31.0
# safetensors==0.3.1
# wandb==0.12.21
# xformers==0.0.21

# # same
# torch==2.0.1
# torchvision==0.15.2
# diffusers==0.20.2
# tqdm==4.64.1

# # different
# accelerate==0.16.0
# transformers==4.27.4
# matplotlib==3.6.2
# opencv_python==4.5.5.64

bitsandbytes==0.41.1
dadaptation==3.1
diffusers==0.20.2
ipython==8.7.0
lion_pytorch==0.1.2
lpips==0.1.4
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.5.5.64
opencv_python_headless==4.7.0.68
pandas==1.5.2
Pillow==10.1.0
prodigyopt==1.0
pydantic==2.6.3
PyYAML==6.0.1
Requests==2.31.0
safetensors==0.3.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
transformers==4.27.4
wandb==0.12.21
xformers==0.0.21
accelerate==0.16.0
